refactor: refactor order by feature with helper class
